refactor: split client responsibilities
Class ConnectionProxy represents now an intermediate level between the Client and the TCP socket

Class ConnectionProxyPool handles the ConnectionProxy instances for resource sharing
