+++ src/util.ts
@@ -16,7 +16,7 @@ export function range(count: number): number[] {
 export const DEFAULT_FRAME_MAX = 1048576
 export const DEFAULT_UNLIMITED_FRAME_MAX = 0
 export const REQUIRED_MANAGEMENT_VERSION = "3.13.0"
-export const getMaxSharedClientInstances = () => {
+export const getMaxSharedConnectionInstances = () => {
   return +(process.env.MAX_SHARED_CLIENT_INSTANCES ?? 100)
 }
